+++ ceammc/CHANGELOG.md
@@ -25,6 +25,11 @@
   - conv.lin2curve: list support added
   - ui.slider2d: mouse wheel support added: (move x-axis when Alt pressed)
   - ui.slider2d alias added: ui.s2d
+- TCL editor added for objects:
+  - data.list
+  - local.list
+  - global.list
+  - data.set
 
 ### Fixed:
 - system.cursor no normalization in subpatch